[pipelines] Remove missed automatic weight casting
This was missed in #60201 that tried to disable automatic weight casting
on load.
As part of this, alternate encodings and automatic weight casting
between bfloat16 and float32 was removed.
While it is nice to be able to run models in a different dtype, this
feature was very ad-hoc and makes it harder to get config correct.

If we want a feature similar to alternative encoding, we should make it
explicitly supported in the graph. Then weights can be cast once on
initialization. Otherwise, out of graph manual solution are fine for
now.

This required changes a few tests from SmolLM2 (bfloat16) to SmolLM
(float32).

MAX_INTEGRATION_TESTS_REV_ID: 543c532ce6b9e64a12ba0d855199ece2a475f10a

+++ max/tests/integration/API/python/test_config_pure.py
@@ -259,7 +259,7 @@ def test_config__validates_supported_device():
         config = PipelineConfig(
             model_path="trl-internal-testing/tiny-random-LlamaForCausalLM",
             device_specs=[DeviceSpec.accelerator()],
-            quantization_encoding=SupportedEncoding.bfloat16,
+            quantization_encoding=SupportedEncoding.float32,
             max_length=1,
         )
